Models, API functiaons, utilities and Admin UI
This code is for push and token auth to Symantec VIP along with
the admin support .

The utilities do the heavy lifting while communication with Symantecs VIP
endpoints is managed from the api file.

SmsDevice and CallDevice are commented out; I don't have access to an instance
of VIP against which to test them.