Move poll settings in to credential model
This is currently the only place these settings are used, and I want to delete
settings.py. If needed, these values can e externalised in some way (eg,
testing and setting a default if unavailable).

I'm taking this chance to remove the import in to utils, where the relevant
code used to live.

I'm going to try and keep 'that helper setup stuff' separate and just have the
queries to VIP in the API file.

This commit bundles a new docstring for make_request_id and a couple of
whitespace fixes lower in the file.