Remove 'return' from credential creation loop
This bug is annoyingly subtle when testing against data which only has a single
credential in it!
The code was returning after a successful save of the first device in a loop
rather than returning after all devices had been created and saved.

I found I wasn't actually looping the records correctly, I've added a test +
data to keep track of it.

For the multi credential user records I now have XML (from the API) and json
(from zeep) stored.