Bring views over from other project
The project I'm using this library in has a better version of some views so
I've upstreamed them here. I'm hoping there will be a couple of others to come.

Using reverse() while urls.py was being processed was causing a loop. Changing
to reverse_lazy defers the parameters check and the file can be parsed again.