Enforce user logins via OTP in middleware
Django-OTP doesn't support an optional second factor login (like
if_configured=True) for its authentication forms.

This middleware is provided for when you want people with OTP configured to
always use it, and those without it configured to never use it.

This effectively reverts and replaces the previous commit "Add VIP component of
global OTP logins".